requirements.txt
wheel
aiohttp
aiofiles
Pillow